#9 working on images size (#17)
* #9 working on images size

* routing naming reviewed

* routing naming reviewed

* #9 image size settled. The error message does not appear if image too big. Due to FileType ?

* picture to tricks added

* add video url for media. WIP

+++ src/EventSubscriber/AdminTrickSubscriber.php
@@ -20,7 +20,7 @@ class AdminTrickSubscriber implements EventSubscriberInterface
         return [
             AdminUploadTrickImageEvent::NAME => [
                 ['upload', 10],
-                ['remove', 20]
+                ['remove', 20],
             ],
         ];
     }
@@ -29,13 +29,19 @@ class AdminTrickSubscriber implements EventSubscriberInterface
     {
        $trick = $event->getTrick();
        $trick->setImage($this->uploader->upload($trick->getImageUpload()));
+
     }
 
     public function remove(AdminUploadTrickImageEvent $event)
     {
         $trick = $event->getTrick();
         $file = $trick->getImage();
-        $file_path = $this->uploader->getTargetDirectory().'/'.$file;
-        if(file_exists($file_path)) unlink($file_path);
+
+        if(isset($file))
+        {
+            $file_path = $this->uploader->getTargetDirectory().'/'.$file;
+            if(file_exists($file_path)) unlink($file_path);
+        }
+
     }
 }
